Added uuid and a multivector for storing components in entity

+++ RockEngine/src/RockEngine/ImGui/ImGuiLayer.cpp
@@ -81,7 +81,7 @@ namespace RockEngine {
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2(static_cast<float>(app.GetWindow().GetWidth()), static_cast<float>(app.GetWindow().GetHeight()));
 
 		// Rendering
 		ImGui::Render();
